+++ src/strategies/default.zsh
@@ -7,9 +7,19 @@
 #
 
 _zsh_autosuggest_strategy_default() {
-	local prefix="$1"
+	# Reset options to defaults and enable LOCAL_OPTIONS
+	emulate -L zsh
+
+	# Enable globbing flags so that we can use (#m)
+	setopt EXTENDED_GLOB
+
+	# Escape backslashes and all of the glob operators so we can use
+	# this string as a pattern to search the $history associative array.
+	# - (#m) globbing flag enables setting references for match data
+	# TODO: Use (b) flag when we can drop support for zsh older than v5.0.8
+	local prefix="${1//(#m)[\\*?[\]<>()|^~#]/\\$MATCH}"
 
 	# Get the history items that match
 	# - (r) subscript flag makes the pattern match on values
-	typeset -g suggestion="${history[(r)${(b)prefix}*]}"
+	typeset -g suggestion="${history[(r)${prefix}*]}"
 }